feat(VsButton): rename icon file
